Seperate file for Player class
The game.py file should bring all the components of the game together.
